+++ src/widget-stats.js
@@ -23,22 +23,24 @@ const WidgetStats = ({ title, value, prev, units, children, trendInfo }) => {
   const classes = useStyles()
   const { isTrendPercentage, upIsGreen, comparedTo, up, down } = trendInfo
   const trendColours = upIsGreen ? ['green', 'red'] : ['red', 'green']
-  const trend = isTrendPercentage
-    ? Math.round(((value - prev) / prev) * 100).toLocaleString()
-    : Math.round(value - prev).toLocaleString()
+  const trend = ((isTrendPercentage
+    ? Math.round(((value - prev) / prev) * 100)
+    : Math.round(value - prev)) || 0).toLocaleString()
 
   return (
-    <Paper className={classes.paper} variant="outlined">
-      <Typography className={classes.title} variant="subtitle2" gutterBottom>
+    <Paper className={classes.paper} variant='outlined'>
+      <Typography className={classes.title} variant='subtitle2' gutterBottom>
         {title}
       </Typography>
       <Typography
-        variant="h5"
+        variant='h5'
         gutterBottom
-      >{`${value.toLocaleString()} ${units}`}</Typography>
-      {children && <Typography variant="body1">{children}</Typography>}
-      {prev && (
-        <Typography variant="body1">
+      >
+        {`${value.toLocaleString()} ${units}`}
+      </Typography>
+      {children && <Typography variant='body1'>{children}</Typography>}
+      {prev !== null && (
+        <Typography variant='body1'>
           {trend > 0 ? (
             <TrendingUpRoundedIcon
               style={{ paddingRight: '8px', color: trendColours[0] }}
@@ -63,7 +65,7 @@ WidgetStats.propTypes = {
   value: PropTypes.number,
   prev: PropTypes.number,
   trendInfo: PropTypes.object,
-  children: PropTypes.object,
+  children: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
   units: PropTypes.string,
 }